[#9]Ao se cadastrar está passando senha menor que 8 ou maior que 12 [#6]Campos numéricos com requisito valores maiores que 0 estão sem validação [correção][javascript]tratamento para valor inexistente 'checked'

+++ db/migrate/00000000000001_create_sites.rb
@@ -14,7 +14,7 @@ class CreateSites < ActiveRecord::Migration[7.1]
         coordenadalt NUMERIC(10, 8),
         coordenadalg NUMERIC(10, 8),
 
-        velocidade_contratada INTEGER NOT NULL,
+        velocidade_contratada NUMERIC NOT NULL,
         codigo_tipo_link INTEGER,
         sla NUMERIC(3, 1),
         valor_mensal VARCHAR(14),
